Task 1 - Prediction using Supervised ML
I have successfully completed #Task1 #GRIPDEC21 #GRIPDECEMBER21 . In the domain of Data Science & Business Analytics. I am very thankful to #TheSparksFoundation for providing me with this opportunity.

In this task, we will be using the Python Scikit-Learn library for machine learning and can be used to implement regression functions.

In this regression task, we will predict the percentage of marks that a student is expected to score based upon the number of hours they studied. This is a simple linear regression task as it involves just two variables.

To Predict: What will be the predicted score if a student studies for 9.25 hrs/ day?

Language Used: Python3 
IDE: Jupyter Notebook 
Dataset URL: http://bit.ly/w-data